src/web_requirements.txt
python-dotenv
fastapi
uvicorn
mysqlclient
pillow
requests
